@@ -4,20 +4,22 @@ version: 0.0.2
 homepage: https://sametates.dev/
 
 environment:
-  sdk: '>=2.19.6 <3.0.0'
+  sdk: '>=3.0.0 <4.0.0'
   flutter: ">=1.17.0"
 
 dependencies:
-  connectivity_plus: ^4.0.0
+  connectivity_plus: ^4.0.1
   nested: ^1.0.0
-  device_info_plus: ^9.0.0
+  device_info_plus: ^9.0.1
+  encrypt: ^5.0.1
+
   flutter:
     sdk: flutter
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_lints: ^2.0.0
+  flutter_lints: ^2.0.1
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
